==> Lessons/04_Lab_DataWrangling_EM.Rmd <==
---
title: "Lab 4: Data Wrangling"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2021"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1- Answer questions on M3/A3
2- Answer questions on M4
3- Practice wrangling datasets with dplyr functions

## Set up your session

Today we will work with a dataset from the [North Temperate Lakes Long-Term Ecological Research Station](https://lter.limnology.wisc.edu/about/overview). The NTL-LTER is located in the boreal zone in northern Wisconsin, USA. We will use the [chemical and physical limnology dataset](https://lter.limnology.wisc.edu/content/cascade-project-north-temperate-lakes-lter-core-data-physical-and-chemical-limnology-1984), running from 1984-2016. 

Opening discussion: why might we be interested in long-term observations of temperature, oxygen, and light in lakes?

> Add notes here: 

```{r, message = FALSE}
getwd()
#install.packages(tidyverse)
library(tidyverse)
#install.packages(lubridate)
library(lubridate)
NTL.phys.data <- read.csv("./Data/Raw/NTL-LTER_Lake_ChemistryPhysics_Raw.csv", stringsAsFactors = TRUE)

class(NTL.phys.data$sampledate)
# Format sampledate as date
NTL.phys.data$sampledate <- as.Date(NTL.phys.data$sampledate, format = "%m/%d/%y")
```


## Filter

Filtering allows us to choose certain rows (observations) in our dataset.

```{r}
class(NTL.phys.data$lakeid)
class(NTL.phys.data$depth)

# dplyr filtering - choosing data where depth =0, aka data from the water column surface
NTL.phys.data.surface <- filter(NTL.phys.data, depth == 0)

# Choose multiple conditions to filter
summary(NTL.phys.data$lakename)
#summary(NTL.phys.data$lakename, maxsum=3) would give us 3 lake names w most occurrrences in dataframe
#choose rows only that has Paul Lake or Peter Lake in the lake name column
NTL.phys.data.PeterPaul <- filter(NTL.phys.data, lakename %in% c("Paul Lake", "Peter Lake"))

# Choose a range of conditions of a numeric or integer variable
summary(NTL.phys.data$daynum)
NTL.phys.data.JunethruOctober <- filter(NTL.phys.data, daynum %in% c(152:304))

# Exercise 1: 
# filter NTL.phys.data for the year 1999
# what code do you need to use, based on the class of the variable?
class(NTL.phys.data$year4)
NTL.phys.data.99 <- filter(NTL.phys.data, year4 == 1999)

# Exercise 2: 
# filter NTL.phys.data for Tuesday Lake from 1990 through 1999.
class(NTL.phys.data$lakename)
NTL.phys.data.Tues.90s <- NTL.phys.data %>%
  filter(year4 %in% c(1990:1999)) %>%
  filter(lakename == "Tuesday Lake")
#== for one element
#%in% for >1 element you're looking for
```
Question: Why don't we filter using row numbers?

> Answer: You can't be sure what order your rows are in unless you look at them all one by one.

## Pipes

Pipe is another method to wrangle datasets that looks cleaner and is easier to read.  We designate a pipe with `%>%`. A good way to think about the function of a pipe is with the word "then." 

Let's say we want to take our raw dataset (NTL.phys.data), *then* filter the data for Peter and Paul lakes, *then* select temperature and observation information, and *then* add a column for temperature in Fahrenheit: 

```{r}
NTL.phys.data.processed <- 
  NTL.phys.data %>%
  filter(lakename == "Paul Lake" | lakename == "Peter Lake") %>%
  select(lakename, sampledate:temperature_C) %>%
  mutate(temperature_F = (temperature_C*9/5) + 32)

#Exercise 3: Using a pipe filter NTL.phys.data for Tuesday Lake from 1990 through 1999 only for July 
class(NTL.phys.data$sampledate)

NTL.phys.data.date <- separate(NTL.phys.data, sampledate, c("Y", "m", "d"))
class(NTL.phys.data.date$m)
NTL.phys.data.date$m <- as.numeric(NTL.phys.data.date$m)
NTL.phys.data.Tues.7.90s <- NTL.phys.data.date %>%
  filter(lakename == "Tuesday Lake") %>% 
  filter(year4 %in% c(1990:1999)) %>%
  filter(m == 7)




#Exercise 4: Using the data from part 3, a pipe and the summarise() function find the mean surface temperature (hints: you will need to add another filter for depth==0). Make sure you eliminate NAs before computing the means
NTL.phys.data.Tues.7.90s.surf <- NTL.phys.data.Tues.7.90s %>%
  filter(depth == 0) %>%
  drop_na(temperature_C) %>%
  summarise(meantemp = mean(temperature_C))
  
```

## Gather and Spread

For gather we will use `pivot_longer` and for spread we will use `pivot_wider`.

```{r}
#Exercise 5: gather irradiance data (measured in the water column and measured on the deck of the sampling boat) into one column using pivot_longer
NTL.phys.data.gathered <- pivot_longer(NTL.phys.data, irradianceWater:irradianceDeck, names_to = "irradiance", values_to = "radiation")


#Exercise 6: spread temperatureC into more than one column based on the depth
NTL.phys.data.spread <- pivot_wider(NTL.phys.data, names_from = "depth", values_from = "temperature_C")
#adds 1 column for each depth with values in column for temperature

```
