==> Lessons/05_Part1_DataVisualizationEM.Rmd <==
---
title: "5: Part 1 - Data Visualization Basics"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2021"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Perform simple data visualizations in the R package `ggplot`
2. Develop skills to adjust aesthetics and layers in graphs
3. Apply a decision tree framework for appropriate graphing methods

## Opening discussion
Effective data visualization depends on purposeful choices about graph types. The ideal graph type depends on the type of data and the message the visualizer desires to communicate. The best visualizations are clear and simple. A good resource for data visualization is [Data to Viz](https://www.data-to-viz.com/), which includes both a decision tree for visualization types and explanation pages for each type of data, including links to R resources to create them. Take a few minutes to explore this website.

## Set Up

```{r, message = FALSE, warning = FALSE}
getwd()
library(tidyverse)
#install.packages("ggridges")
library(ggridges)
#ridges helps you view ridge plots, that look at distribution of a variable over time

PeterPaul.chem.nutrients <- 
  read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv", stringsAsFactors = TRUE)
PeterPaul.chem.nutrients.gathered <-
  read.csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv", stringsAsFactors = TRUE)
EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1718_Processed.csv", stringsAsFactors = TRUE)
#we previously saved these processed data files from other sessions - changed to 1718 bc that's what A04 said to do ^

EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")
PeterPaul.chem.nutrients.gathered$sampledate <- as.Date(PeterPaul.chem.nutrients.gathered$sampledate, format = "%Y-%m-%d")
#changing dates from factors to dates
```

## ggplot

ggplot, called from the package `ggplot2`, is a graphing and image generation tool in R. This package is part of tidyverse. While base R has graphing capabilities, ggplot has the capacity for a wider range and more sophisticated options for graphing. 
ggplot has only a few rules: 

* The first line of ggplot code always starts with `ggplot()`
* A data frame must be specified within the `ggplot()` function. Additional datasets can be specified in subsequent layers.
* Aesthetics must be specified, most commonly x and y variables but including others. Aesthetics can be specified in the `ggplot()` function or in subsequent layers.
* Additional layers must be specified to fill the plot. 

### Geoms

Here are some commonly used layers for plotting in ggplot: 

* geom_bar
* geom_histogram
* geom_freqpoly
* geom_boxplot
* geom_violin
* geom_dotplot
* geom_density_ridges
* geom_point
* geom_errorbar
* geom_smooth
* geom_line
* geom_area
* geom_abline (plus geom_hline and geom_vline)
* geom_text

### Aesthetics
Here are some commonly used aesthetic types that can be manipulated in ggplot: 

* color
* fill
* shape
* size
* transparency

### Plotting continuous variables over time: Scatterplot and Line Plot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
#above we are specifying the dimensions of the figure (plot) that comes from this box of code
# Scatterplot
ggplot(EPAair, aes(x = Date, y = Ozone)) + 
  geom_point()
#basic scatterplot of date v air ozone level
#geom_point = you get 1 pt per data cell

O3plot <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone))
print(O3plot)
#same plot as above but generated differently: creating an object that we can add layers to below, then print it each time we add layers
#notice we specified x and y on the second line rather than first but got the same plot

# Fix this code
O3plot2 <- ggplot(EPAair) +
  #geom_point(aes(x = Date, y = Ozone, color = "blue"))
  #running above # code, color is orange and R adds a legend calling all points 'blue' --> we specified blue inside the x, y = () and this added a layer of points ratehr than just aesthetics
  geom_point(aes(x = Date, y = Ozone), color = "blue")
print(O3plot2)
#referring to above plot object, O3plot2, but making points blue

# Add additional variables
# How could you automatically assign a marker color to a variable - ggplot automatically does this. 
PMplot <- 
  ggplot(EPAair, aes(x = Month, y = PM2.5, shape = as.factor(Year), color = Site.Name)) +
  geom_point()
print(PMplot)
#telling R to auto assign diff shapes based on Year and color based on Site Name. now x = month, not year. but visually this doesn't tell us much. so below we try to adjust it by separating by facet:

#if your plot looks squished you may need to change the dimensions that you specified for your figure in the r chunk 

# Separate plot with facets
PMplot.faceted <-
  ggplot(EPAair, aes(x = Month, y = PM2.5, shape = as.factor(Year))) +
  geom_point() +
  facet_wrap(vars(Site.Name), nrow = 3)
print(PMplot.faceted)
#removing color part, separating by facet [here, specified as site name] --> instead of having site vary by color, we create one little plot per site
#chose 3 rows bc we have 12 sites
#still separated by month bc we want to see how PM and O3 are changing throughout the year

# Filter dataset within plot building and facet by multiple variables
PMplot.faceted2 <- 
  ggplot(subset(EPAair, Site.Name == "Clemmons Middle" | Site.Name == "Leggett" |
                  Site.Name == "Bryson City"), 
         aes(x = Month, y = PM2.5)) + 
  geom_point() +
  facet_grid(Site.Name ~ Year) #col = yr, row = 3 sites we selected (for whole plot scheme, not individual facet plots)
print(PMplot.faceted2)
#above: if we're not interested in viewing all sites, we can specify specific sites using subset fxn
#| means or, so R will choose all these sites

# Plot true time series with geom_line
PMplot.line <- 
  ggplot(subset(EPAair, Site.Name == "Leggett"), 
         aes(x = Date, y = PM2.5)) +
  geom_line()
print(PMplot.line)
#above we use a line rather than point plot - connecting the points.
#this is a time series plot

#all above plots used 1 cont variable and added layers w other variables

```


### Plotting the relationship between two continuous variables: Scatterplot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
lightvsDO <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point()
print(lightvsDO)
#above is a scatterplot w 2 variables (versus 1 var and date)
#most data is bunched in bottom left - want to adjust via code below to account for the couple outliers we have

# Adjust axes
lightvsDOfixed <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point() +
  xlim(0, 250) +
  ylim(0, 20)
print(lightvsDOfixed)
#adding limitations to x and y axis to make visualization more useful

# Depth in the fields of limnology and oceanography is on a reverse scale
tempvsdepth <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth)) +
  #ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth, color = daynum)) +
  geom_point() +
  scale_y_reverse()
print(tempvsdepth)
#scale y reverse will switch your y scale so it goes big to small. we do this here bc we are plotting depth on y - so want to go from shallow water at top of graph to deep water at bottom for visualization
#see that as we get deeper, temp is lowered

NvsP <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, color = depth)) +
  geom_point() +
  geom_smooth(method = lm) +
  geom_abline(aes(slope = 16, intercept = 0))
print(NvsP)
#categorize the observations by depth via color=depth bc we want to see if relationship of N vs P is affected by depth
#geom_smooth adds in line of best fit
#method = lm --> use linear model to plot line of best fit. blue line is this trendline
#geom_abline adds the black line in which we specify slope and intercept 


```

### Plotting continuous vs. categorical variables
A traditional way to display summary statistics of continuous variables is a bar plot with error bars. Let's explore why this might not be the most effective way to display this type of data. Navigate to the Caveats page on Data to Viz (https://www.data-to-viz.com/caveats.html) and find the page that explores barplots and error bars. 

What might be more effective ways to display the information? Navigate to the boxplots page in the Caveats section to explore further.

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Box and whiskers plot
Nutrientplot3 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_boxplot(aes(color = nutrient)) # Why didn't we use "fill"?
print(Nutrientplot3)
#using gathered dataset so we can see all the nutrients
#separated columns by lake name, did colors based on nutrient type
#fill is the filling of the shape, color is the outline of the shape
#fill keeps lines and dots in boxplots black so not useful
#can't see some of the boxes - hard to make assumptions and visualize things

# Dot plot
Nutrientplot4 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_dotplot(aes(color = nutrient, fill = nutrient), binaxis = "y", binwidth = 1, 
               stackdir = "center", position = "dodge", dotsize = 2) #
print(Nutrientplot4)
#above changing contour AND filling based on nutrient
#using dotplot rather than boxplot to see if we can better visualize some of the lake data that we couldn't see the boxes for
#binaxis = y means we stack observations based on y
#stackdir = center means we stack points based on center position
#position = dodge lets us avoid dots overlaying each other - they will go to side of each other rather than overlay
#y axis scale makes it hard for us to visualize data characteristics though

# Violin plot
Nutrientplot5 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_violin(aes(color = nutrient)) #
print(Nutrientplot5)
#now trying a violin to see if that makes things better visually
#violin is like a rotated density plot that repeats on both sides

# Frequency polygons
# Using a tidy dataset
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients) +
  geom_freqpoly(aes(x = tn_ug), color = "darkred") +
  geom_freqpoly(aes(x = tp_ug), color = "darkblue") +
  geom_freqpoly(aes(x = nh34), color = "blue") +
  geom_freqpoly(aes(x = no23), color = "royalblue") +
  geom_freqpoly(aes(x = po4), color = "red") 
print(Nutrientplot6)
#making a different frequency polygon for each nutrient and deisgnating them as diff colors.
#issues: no legend so we don't know what diff colors mean. x axis only added label for first freq polygon - others we added are just layers on that graph. this is also kind of text heavy in the code. easier way to accomplish this below by using the gathered dataset ratehr than non-gathered (nutrients in sep columns):

# Using a gathered dataset
Nutrientplot7 <-   
  ggplot(PeterPaul.chem.nutrients.gathered) +
  geom_freqpoly(aes(x = concentration, color = nutrient))
print(Nutrientplot7)
#when we use color inside aesthetics, we get a legend automatically.

# Frequency polygons have the risk of becoming spaghetti plots. 
# See <https://www.data-to-viz.com/caveat/spaghetti.html> for more info.
#^if you have too many variables, it gets too crowded and complicated to interpret

# Ridgeline plot
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(y = nutrient, x = concentration)) +
  geom_density_ridges(aes(fill = lakename), alpha = 0.5) #
print(Nutrientplot6)
#fill fills the density below the curve of the plot with a color

```
