==> Lessons/06_Part2_GLMs_EM.Rmd <==
---
title: "6: Part 2 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2021"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
2. Apply special cases of the GLM (ANOVA) to real datasets
3. Interpret and report the results of ANOVA in publication-style formats

## Set up
```{r, message = FALSE}
getwd()
library(ggplot2)
library(tidyverse)
install.packages("agricolae")
install.packages("htmltools")
library(htmltools)
library(agricolae)   #you may need to install/update package "htmltools" before installing and loading agricolae

Litter <- read.csv("./Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)

# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

### One-way ANOVA
A one-way ANOVA is used when you have one categorical explanatory variable with more than two categories. In R, we can run the model with the function `lm` or `aov`, the latter of which will allow us to run post-hoc tests to determine pairwise differences.

Does total litter biomass differ among sites at the Niwot Ridge LTER?
```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>%
  summarise(dryMass = sum(dryMass))
#will reject or not reject H0 based on anova output. rejecting tells you that at least one pop's mean is not equal. BUT anova doesn't tell you which one - have to do post-hoc tests for this. 

summary(Litter.Totals)
#check all value of plotID
#want to know if high variance we see in mean is due to site location
summary(Litter.Totals$plotID)

#checking assumptions:
# Test for normality. 
# Note: the important assumption for GLMs is normality of residuals, 
# not necessarily normality of raw data. See below when we plot the model.
# shapiro.test() performs the Shapiro-Wilk test of the null that data follows a normal distribution

#filtering drymass by sites:
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_040"]) #reject H0
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_041"]) #reject H0
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_046"]) #keep H0
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_047"]) #reject H0
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_051"]) #reject H0
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_057"]) #keep H0
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_058"]) #reject H0
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_061"]) #reject H0
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_062"]) #reject
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_063"]) #reject
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_064"]) #reject
shapiro.test(Litter.Totals$dryMass[Litter.Totals$plotID == "NIWO_067"]) #reject

#results: reject null in all except two: NIWO_057 and NIWO_046

#can use qqplot to check for normality
qqnorm(Litter.Totals$dryMass); qqline(Litter.Totals$dryMass)

# Test for equal variance
# ANOVA is robust against departures from equal variance.
# bartlett.test() performs Bartlett's test of the null that the variances in each of the groups (samples) are the same.
bartlett.test(Litter.Totals$dryMass ~ Litter.Totals$plotID)

#results: reject null i.e. variances are not equal

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova)

#first row of results is the factor, second is the residuals
#mean sq is sum sq divided by degrees of freedom
#results: reject null hypothesis i.e. difference between a pair of group means is statiscally significant

#we know not all means are the same from this but don't know which means are diff

# Format ANOVA as lm - main difference is output summary table. above is traditional anova but below has 1 row per level of categorical var.
Litter.Totals.anova2 <- lm(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova2)


# Checking model fit and assumptions
# ANOVA is robust against departures from normality.
plot(Litter.Totals.anova2)

# Post-hoc test
# TukeyHSD() computes Tukey Honest Significant Differences
TukeyHSD(Litter.Totals.anova)

# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE)
Litter.Totals.groups
#letters in groups help you categorize your dataset and find which levels share the same mean and could be in the same treatment group
#everything with an 'a' has statistically similar means. could filter dataset to see only the sites that contain an 'a' - results of running this new filtered dataset in anova will show that their means are not significantly different. if a site has 'abcd' it could be grouped (bc of similar means) with other sites containing a, b, c, and/or d (basically, you fit the equal means assumption of the anova if you group levels together based on letters they share.) <- think of it like a venn diagram

# Graph the results
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot)
 
```

What information does the output give us? How might we report this information in a report?

> need to reorder sites so they're in the same order as the groups list from HSD test so we can better see changes in mean

## TWO-WAY ANOVA
### Main effects
A two-way ANOVA allows us to examine the effects of two categorical explanatory variables on a continuous response variable. What if we wanted to know how litter masses varied across functional groups and land cover class?

```{r}
# Format as aov
Litter.anova.2way <- aov(data = Litter, dryMass ~ functionalGroup + nlcdClass)
summary(Litter.anova.2way)
#df for residuals is associated w number of observations we have (total obs - other dfs - 1)

# Format as lm
Litter.anova.2way2 <- lm(data = Litter, dryMass ~ functionalGroup + nlcdClass)
summary(Litter.anova.2way2)
#gives same overall results as aov - both point to differences in means.

# Run a post-hoc test for pairwise differences
TukeyHSD(Litter.anova.2way)

```

### Interaction effects
We may expect the effects of functional group and land cover to be dependent on each other. For instance, we may expect needle biomass to be larger in evergreen forests than in other land types. In this case, we might choose to run an interaction effects two-way ANOVA, which will examine the individual effects of the explanatory variables as well as the interaction of the explanatory variables.

The output gives test statistics for each explanatory variable as well as the interaction effect of the explanatory variables. If the p-value for the interaction effect is less than 0.05, then we would consider the interaction among the explanatory variables to be significant.

```{r}
# Format as aov
Litter.anova.2way3 <- aov(data = Litter, dryMass ~ functionalGroup * nlcdClass)
summary(Litter.anova.2way3)
#p value <0.05 suggests an interaction bw vars is sig
#suggests there is a significant interaction bw functional group and NLCD class

# Format as lm
Litter.anova.2way4 <- lm(data = Litter, dryMass ~ functionalGroup * nlcdClass)
summary(Litter.anova.2way4)

```

If the interaction is significant, we interpret pairwise differences for the interaction. If the interaction is not significant, we interpret differences for the main effects only.
```{r}
# Run a post-hoc test for pairwise differences
TukeyHSD(Litter.anova.2way3)

#create treatment groups via the interaction that's significant:
Litter.interaction <- with(Litter, interaction(functionalGroup, nlcdClass))
#above is a factor w 24 levels

#Litter.test <- cbind(Litter,Litter.interaction)

Litter.anova.2way5 <- aov(data = Litter, dryMass ~ Litter.interaction)
#one way anova but the explan is a result of an interaction bw 2 cat variables

Litter.groups <- HSD.test(Litter.anova.2way5, "Litter.interaction", group = TRUE)
Litter.groups

# Plot the results
# Edit this graph to make it more visually appealing.
# Add letters to represent statistical groupings.
Litter.anova.plot <- ggplot(Litter, aes(y = dryMass, x = nlcdClass, color = functionalGroup)) +
  geom_boxplot()
print(Litter.anova.plot)


#edit graph to make it more visually appealing and add letters like in 1 way anova to represent statistical grouping
```
